requirements.txt
flask==3.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
openai==1.82.0
stripe==12.1.0
Pillow==11.1.0
python-dotenv==1.1.0
